refactor to Serial and ThreadedTestCase structs, and move files to dp pkg
