+++ npm/pkg/dataplane/types_windows_test.go
@@ -1,31 +1,116 @@
-package e2e
+package dataplane
 
 import (
 	"fmt"
+	"time"
 
+	"github.com/Azure/azure-container-networking/network/hnswrapper"
 	"github.com/Azure/azure-container-networking/npm/pkg/controlplane/translation"
-	"github.com/Azure/azure-container-networking/npm/pkg/dataplane"
 	"github.com/Azure/azure-container-networking/npm/pkg/dataplane/ipsets"
+	dptestutils "github.com/Azure/azure-container-networking/npm/pkg/dataplane/testutils"
+	"github.com/Microsoft/hcsshim/hcn"
 	"github.com/pkg/errors"
-
 	networkingv1 "k8s.io/api/networking/v1"
 )
 
+type Tag string
+
+type SerialTestCase struct {
+	Description string
+	Actions     []*Action
+	*TestCaseMetadata
+}
+
+type ThreadedTestCase struct {
+	Description string
+	HNSLatency  time.Duration
+	Threads     map[string][]*Action
+	*TestCaseMetadata
+}
+
+type TestCaseMetadata struct {
+	Tags                 []Tag
+	InitialEndpoints     []*hcn.HostComputeEndpoint
+	DpCfg                *Config
+	ExpectedSetPolicies  []*hcn.SetPolicySetting
+	ExpectedEnpdointACLs map[string][]*hnswrapper.FakeEndpointPolicy
+}
+
+// Action represents a single action (either an HNSAction or a DPAction).
+// Exactly one of HNSAction or DPAction should be non-nil.
+type Action struct {
+	HNSAction
+	DPAction
+}
+
+type HNSAction interface {
+	Do(hns *hnswrapper.Hnsv2wrapperFake) error
+}
+
+type EndpointCreateAction struct {
+	ID string
+	IP string
+}
+
+func CreateEndpoint(id, ip string) *Action {
+	return &Action{
+		HNSAction: &EndpointCreateAction{
+			ID: id,
+			IP: ip,
+		},
+	}
+}
+
+func (e *EndpointCreateAction) Do(hns *hnswrapper.Hnsv2wrapperFake) error {
+	ep := dptestutils.Endpoint(e.ID, e.IP)
+	_, err := hns.CreateEndpoint(ep)
+	if err != nil {
+		return errors.Wrapf(err, "[EndpointCreateAction] failed to create endpoint. ep: [%+v]", ep)
+	}
+	return nil
+}
+
+type EndpointDeleteAction struct {
+	ID string
+}
+
+func DeleteEndpoint(id string) *Action {
+	return &Action{
+		HNSAction: &EndpointDeleteAction{
+			ID: id,
+		},
+	}
+}
+
+func (e *EndpointDeleteAction) Do(hns *hnswrapper.Hnsv2wrapperFake) error {
+	ep := &hcn.HostComputeEndpoint{
+		Id: e.ID,
+	}
+	if err := hns.DeleteEndpoint(ep); err != nil {
+		return errors.Wrapf(err, "[EndpointDeleteAction] failed to delete endpoint. ep: [%+v]", ep)
+	}
+	return nil
+}
+
+type DPAction interface {
+	Do(dp *DataPlane) error
+}
+
 type PodCreateAction struct {
-	Pod    *dataplane.PodMetadata
+	Pod    *PodMetadata
 	Labels map[string]string
 }
 
 func CreatePod(namespace, name, ip, node string, labels map[string]string) *Action {
 	return &Action{
 		DPAction: &PodCreateAction{
-			Pod:    dataplane.NewPodMetadata(fmt.Sprintf("%s/%s", namespace, name), ip, node),
+			Pod:    NewPodMetadata(fmt.Sprintf("%s/%s", namespace, name), ip, node),
 			Labels: labels,
 		},
 	}
 }
 
-func (p *PodCreateAction) Do(dp *dataplane.DataPlane) error {
+func (p *PodCreateAction) Do(dp *DataPlane) error {
 	context := fmt.Sprintf("create context: [pod: %+v. labels: %+v]", p.Pod, p.Labels)
 
 	nsIPSet := []*ipsets.IPSetMetadata{ipsets.NewIPSetMetadata(p.Pod.Namespace(), ipsets.Namespace)}
@@ -62,7 +147,7 @@ func ApplyDP() *Action {
 	}
 }
 
-func (*ApplyDPAction) Do(dp *dataplane.DataPlane) error {
+func (*ApplyDPAction) Do(dp *DataPlane) error {
 	if err := dp.ApplyDataPlane(); err != nil {
 		return errors.Wrapf(err, "[ApplyDPAction] failed to apply")
 	}
@@ -85,7 +170,7 @@ func UpdatePolicy(policy *networkingv1.NetworkPolicy) *Action {
 	}
 }
 
-func (p *PolicyUpdateAction) Do(dp *dataplane.DataPlane) error {
+func (p *PolicyUpdateAction) Do(dp *DataPlane) error {
 	npmNetPol, err := translation.TranslatePolicy(p.Policy)
 	if err != nil {
 		return errors.Wrapf(err, "[PolicyUpdateAction] failed to translate policy with key %s/%s", p.Policy.Namespace, p.Policy.Name)
@@ -115,7 +200,7 @@ func DeletePolicyByObject(policy *networkingv1.NetworkPolicy) *Action {
 	return DeletePolicy(policy.Namespace, policy.Name)
 }
 
-func (p *PolicyDeleteAction) Do(dp *dataplane.DataPlane) error {
+func (p *PolicyDeleteAction) Do(dp *DataPlane) error {
 	policyKey := fmt.Sprintf("%s/%s", p.Namespace, p.Name)
 	if err := dp.RemovePolicy(policyKey); err != nil {
 		return errors.Wrapf(err, "[PolicyDeleteAction] failed to update policy with key %s", policyKey)
